Fix db_lock() prototype.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@1489 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ datastore_db.c
@@ -59,7 +59,7 @@ typedef struct {
 
 static long db_get_dbvalue(void *vhandle, const char *word, dbv_t *val);
 static void db_set_dbvalue(void *vhandle, const char *word, dbv_t *val);
-static int db_lock(int fd, int cmd, int type);
+static int db_lock(int fd, int cmd, short int type);
 
 /* Function definitions */
